pyproject.toml
[tool.poetry]
name = "notesapi"
version = "0.1.0"
description = ""
authors = ["RoieGolst <roiegol9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
pymongo = "^4.8.0"
jsonschema = "^4.23.0"
beanie = "^1.26.0"
motor = "^3.5.1"
pyjwt = "^2.8.0"
cryptography = "^43.0.0"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

